confgen: Develop test for complex autoexec
So far we sucessfully used normal Python3 configparser to read DOSBox'
conf files format.  However it breaks as soon as 'if x == y' statement
appears in the instruction list.

To fix it, we need to subclass configparser.ConfigParser to disable
split by key/value specifically for autoexec section.

This (disabled) test provides basis for implementation.