Added Steam Play to project description
Added Steam Play to project description to indicate clearly that the tool is used as a part of *Steam Play*, paraphrasing what the screenshot below says. Having it in textual description may ease up indexing and finding the project in search engines, too.

Will do the same PR for Luxtorpeda. Project description on GitHub could use similar update as well.